Fix for FindNamespacesForArea
If an area had some routes defined with a namespace and some without
then those without would not be included in the sitemap. Now such a
route leads to an extra empty string being added to the returned
enumerable so the whole assembly is searched as is the case if no routes
with namesapces are found.