Update src/MvcSiteMapProvider/MvcSiteMapProvider/DefaultSiteMapProvider.cs
Hi, first pull request - be gentle!

Fixed a potential bug around finding a root node. Change from Request.RawUrl to Request.Path

The scenario this affects is this:

Site root '/' maps to Controller/Action "Home/Index" that in turn performs a _Transfer_ to a controller/Action deeper in the site
e.g. /Products/Browse

With the fix from issue #71 ("Checking for matching route values when finding the current node")
This now breaks (FindSiteMapNode returns null) as RawUrl in this case is "/", which resolves to the node at "Home/Index" but the route data is for a deeper node Controller = "Products", Action="Browse"

Using Path correctly returns "/Products/Browse" which resolves the correct node "Products/Browse" that matches the specifed routeData Controller="Products", Action ="Browse"

Does this make sense?