Update src/MvcSiteMapProvider/MvcSiteMapProvider/DefaultSiteMapProvider.cs
Moved the BuildSiteMap caching into the finally block. This allows exceptions to be caught and, if required by the application, for the sitemap to be refreshed.

Currently an exception thrown while building (eg in a dynamicnodeprovider) will result in a sitemap being partially built without the cachekey being added, meaning that the refresh method won't work.