SiteMapNode did not handle multiple sitemaps defined in the web.config.  If you defined the sitemap in code via attributes, there was no way to specify which sitemap the node belonged to.  Added this functionality.  By default if you only have one sitemap, everything works as it did, otherwise there is a new 'SiteMapProviderName' property of a MvcSiteMapNode.
The logic used to determine the area using the namespace was returning areas when it should not have. A namespace of Company.Webs.MainSite.Controllers would return 'Webs' as the area.  According to the default conventions, there is no area here.  The default convention looks like Company.Webs.MainSite.Areas.MyAccount.Controllers where MyAccount is the area.  Adjusted logic to properly parse the area out of namespaces.